ADEN-7004 Move slots setup to slots.js
