+++ app/modules/ads/setup.js
@@ -5,93 +5,9 @@ import slots from './slots';
 import SlotTracker from './tracking/slot-tracker';
 import targeting from './targeting';
 import ViewabilityTracker from './tracking/viewability-tracker';
-import offset from "../../utils/offset";
-
-const MIN_ZEROTH_SECTION_LENGTH = 700;
-const MIN_NUMBER_OF_SECTIONS = 4;
-const PAGE_TYPES = {
-	article: 'a',
-	home: 'h'
-};
-
-function setSlotState(slotName, state) {
-	const slotService = Wikia.adEngine.slotService;
-
-	if (state) {
-		slotService.enable(slotName);
-	} else {
-		slotService.disable(slotName);
-	}
-}
-
-function isTopLeaderboardApplicable() {
-	const context = Wikia.adEngine.context;
-
-	const hasFeaturedVideo = context.get('custom.hasFeaturedVideo'),
-		isHome = context.get('custom.pageType') === 'home',
-		hasPageHeader = !!document.querySelector('.wiki-page-header'),
-		hasPortableInfobox = !!document.querySelector('.portable-infobox');
-
-	return isHome || hasPortableInfobox || (hasPageHeader > 0 && !hasFeaturedVideo);
-}
-
-function isInContentApplicable() {
-	const context = Wikia.adEngine.context;
-
-	if (context.get('custom.pageType') === 'home') {
-		return !!document.querySelector('.curated-content');
-	}
-
-	const firstSection = document.querySelector('.article-content > h2'),
-		firstSectionTop = (
-			firstSection &&
-			offset(firstSection).top
-		) || 0;
-
-	return firstSectionTop > MIN_ZEROTH_SECTION_LENGTH;
-}
-
-function isPrefooterApplicable(isInContentApplicable) {
-	const context = Wikia.adEngine.context;
-
-	if (context.get('custom.pageType') === 'home') {
-		return !!document.querySelector('.trending-articles');
-	}
-
-	const numberOfSections = document.querySelectorAll('.article-content > h2').length,
-		hasArticleFooter = !!document.querySelector('.article-footer');
-
-	return hasArticleFooter && !isInContentApplicable || numberOfSections > MIN_NUMBER_OF_SECTIONS;
-}
-
-function isBottomLeaderboardApplicable() {
-	return !!document.querySelector('.wds-global-footer');
-}
-
-function setupSlotsStates() {
-	const context = Wikia.adEngine.context;
-
-	const incontentState = isInContentApplicable();
-
-	setSlotState('MOBILE_TOP_LEADERBOARD', isTopLeaderboardApplicable());
-	setSlotState('MOBILE_IN_CONTENT', incontentState);
-	setSlotState('MOBILE_PREFOOTER', isPrefooterApplicable(incontentState));
-	setSlotState('BOTTOM_LEADERBOARD', isBottomLeaderboardApplicable());
-	setSlotState('FEATURED', context.get('custom.hasFeaturedVideo'));
-}
-
-function getPageTypeShortcut() {
-	// Global imports:
-	const {context} = window.Wikia.adEngine;
-	// End of imports
-
-	return PAGE_TYPES[context.get('targeting.s2')] || 'x';
-}
 
 function setupPageLevelTargeting(mediaWikiAdsContext) {
-	// Global imports:
 	const {context} = window.Wikia.adEngine;
-	// End of imports
 
 	const pageLevelParams = targeting.getPageLevelTargeting(mediaWikiAdsContext);
 	Object.keys(pageLevelParams).forEach((key) => {
@@ -99,31 +15,13 @@ function setupPageLevelTargeting(mediaWikiAdsContext) {
 	});
 }
 
-function setupSlotIdentificator() {
-	// Global imports:
-	const {context} = window.Wikia.adEngine;
-	// End of imports
-
-	const pageTypeParam = getPageTypeShortcut();
-	const slotsDefinition = context.get('slots');
-
-	// Wikia Page Identificator
-	context.set('targeting.wsi', `mx${pageTypeParam}1`);
-	Object.keys(slotsDefinition).forEach((key) => {
-		const slotParam = slotsDefinition[key].slotShortcut || 'x';
-		context.set(`slots.${key}.targeting.wsi`, `m${slotParam}${pageTypeParam}1`);
-	});
-}
-
 function setupAdContext(adsContext, instantGlobals) {
-	// Global imports:
 	const {context, utils} = window.Wikia.adEngine;
 	const {isProperGeo} = window.Wikia.adProductsGeo;
 
 	function isGeoEnabled(instantGlobalKey) {
 		return isProperGeo(instantGlobals[instantGlobalKey]);
 	}
-	// End of imports
 
 	context.extend(basicContext);
 
@@ -171,14 +69,12 @@ function setupAdContext(adsContext, instantGlobals) {
 	context.set('custom.hasPortableInfobox', adsContext.targeting.hasPortableInfobox);
 	context.set('custom.pageType', adsContext.targeting.pageType);
 
-	setupSlotIdentificator();
-	setupSlotsStates();
+	slots.setupIdentificators();
+	slots.setupStates();
 }
 
 function configure(adsContext, instantGlobals) {
-	// Global imports:
 	const {context} = window.Wikia.adEngine;
-	// End of imports
 
 	setupAdContext(adsContext, instantGlobals);
 	// TODO: run tracking on each pv (result is already stored, just call adBlockDetection.track() method)
@@ -190,9 +86,7 @@ function configure(adsContext, instantGlobals) {
 }
 
 function init() {
-	// Global imports:
 	const {AdEngine} = window.Wikia.adEngine;
-	// End of imports
 
 	const engine = new AdEngine();
